+++ decoder.go
@@ -262,11 +262,46 @@ func (d *decoder) scalar(childKey string, value reflect.Value, def string) error
 	return convert(childKey, &value, val)
 }
 
-// Set value for a sequence type
-// TODO(alsam) We stop populating sequence on a first nil value. Can we do better?
+// Collections of objects can be converted in the following way:
+// 1. Maps can be decoded only to maps
+// 2. Slices with arrays can be decoded one into another.
+func checkCollections(src, dst reflect.Kind) error {
+	err := errors.Errorf("can't convert %q to %q", src, dst)
+	switch dst {
+	case reflect.Map:
+		if src != reflect.Map {
+			return err
+		}
+	case reflect.Array, reflect.Slice:
+		if src != reflect.Array && src != reflect.Slice {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// Set value for a sequence type. Length of the collection is determined by the
+// length of the underlying collection in a provider. It can be augmented
+// further, by overriding values after the end, e.g.
+//
+// 	var x []int
+// 	NewStaticProvider(map[string]interface{}{"a":[]int{0,1}, "a.2":2}.Populate(&x)
+// 	fmt.Println(x)
+//
+// will print:
+// [0 1 2]
 func (d *decoder) sequence(childKey string, value reflect.Value) error {
-	valueType := value.Type()
 	global := d.getGlobalProvider()
+	slice := global.Get(childKey)
+
+	s := slice.Value()
+	sv := reflect.ValueOf(s)
+	if err := checkCollections(sv.Kind(), value.Kind()); err != nil && slice.Value() != nil {
+		return err
+	}
+
+	valueType := value.Type()
 	destSlice := reflect.MakeSlice(valueType, 0, 4)
 
 	// start looking for child values.
@@ -277,7 +312,11 @@ func (d *decoder) sequence(childKey string, value reflect.Value) error {
 		arrayKey := childKey + strconv.Itoa(ai)
 
 		// Iterate until we find first missing value.
-		if v2 := global.Get(arrayKey); v2.HasValue() {
+		if v2 := global.Get(arrayKey); v2.Value() != nil {
+			if err := checkCollections(reflect.TypeOf(v2.value).Kind(), elementType.Kind()); err != nil {
+				return err
+			}
+
 			val := reflect.New(elementType).Elem()
 
 			// Unmarshal current element.
@@ -286,12 +325,18 @@ func (d *decoder) sequence(childKey string, value reflect.Value) error {
 			}
 
 			// Append element to the slice
-			if destSlice.Len() <= ai {
-				destSlice = reflect.Append(destSlice, reflect.Zero(elementType))
-			}
-
+			destSlice = reflect.Append(destSlice, reflect.Zero(elementType))
 			destSlice.Index(ai).Set(val)
 		} else {
+			// Value in the middle was overridden,
+			// but is missing, set it to the zero initialized value.
+			if sv.IsValid() && sv.Len() > ai {
+
+				// Append element to the slice
+				destSlice = reflect.Append(destSlice, reflect.Zero(elementType))
+				continue
+			}
+
 			break
 		}
 	}
@@ -303,10 +348,18 @@ func (d *decoder) sequence(childKey string, value reflect.Value) error {
 	return nil
 }
 
-// Set value for the array type
+// Set value for the array type. If a value for array item is not found,
+// decoder will use a default one if present, or a zero constructed.
 func (d *decoder) array(childKey string, value reflect.Value) error {
-	valueType := value.Type()
 	global := d.getGlobalProvider()
+	ar := global.Get(childKey)
+	a := ar.Value()
+	av := reflect.ValueOf(a)
+	if err := checkCollections(av.Kind(), value.Kind()); err != nil && ar.Value() != nil {
+		return err
+	}
+
+	valueType := value.Type()
 
 	// start looking for child values.
 	elementType := derefType(valueType).Elem()
@@ -316,7 +369,11 @@ func (d *decoder) array(childKey string, value reflect.Value) error {
 		arrayKey := childKey + strconv.Itoa(ai)
 
 		// Iterate until we find first missing value.
-		if v2 := global.Get(arrayKey); v2.HasValue() {
+		if v2 := global.Get(arrayKey); v2.Value() != nil {
+			if err := checkCollections(reflect.TypeOf(v2.value).Kind(), elementType.Kind()); err != nil {
+				return err
+			}
+
 			val := reflect.New(elementType).Elem()
 
 			// Unmarshal current element.
@@ -346,39 +403,39 @@ func (d *decoder) mapping(childKey string, value reflect.Value, def string) erro
 	}
 
 	val := v.Value()
-	destMap := reflect.ValueOf(reflect.MakeMap(valueType).Interface())
-
-	// child yamlNode parsed from yaml file is of type map[interface{}]interface{}
-	// type casting here makes sure that we are iterating over a parsed map.
-	if v, ok := val.(map[interface{}]interface{}); ok {
-		childKey = addSeparator(childKey)
-
-		for key := range v {
-			subKey := fmt.Sprintf("%v", key)
-			if subKey == "" {
-				// We can confuse an empty map key with a root element.
-				return errorWithKey(errors.New("empty map key is ambiguous"), childKey)
-			}
-
-			itemValue := reflect.New(valueType.Elem()).Elem()
-
-			// Try to unmarshal value and save it in the map.
-			if err := d.unmarshal(childKey+subKey, itemValue, def); err != nil {
-				return err
-			}
-
-			//TODO(alsam) do we need non scalar key types?
-			keyVal := reflect.New(value.Type().Key()).Elem()
-			if err := convert(childKey, &keyVal, key); err != nil {
-				return errors.Wrap(err, "key types conversion")
-			}
-
-			destMap.SetMapIndex(keyVal, itemValue)
-		}
-
-		value.Set(destMap)
+	if err := checkCollections(reflect.TypeOf(v.Value()).Kind(), valueType.Kind()); err != nil {
+		return fmt.Errorf("expected map for key %q. actual type: %q", childKey, reflect.TypeOf(val))
 	}
 
+	destMap := reflect.ValueOf(reflect.MakeMap(valueType).Interface())
+
+	childKey = addSeparator(childKey)
+
+	rVal := reflect.ValueOf(val)
+	for _, key := range rVal.MapKeys() {
+		subKey := fmt.Sprintf("%v", key.Interface())
+		if subKey == "" {
+			// We can confuse an empty map key with a root element.
+			return errorWithKey(errors.New("empty map key is ambiguous"), childKey)
+		}
+
+		itemValue := reflect.New(valueType.Elem()).Elem()
+
+		// Try to unmarshal value and save it in the map.
+		if err := d.unmarshal(childKey+subKey, itemValue, def); err != nil {
+			return err
+		}
+
+		//TODO(alsam) do we need non scalar key types?
+		keyVal := reflect.New(value.Type().Key()).Elem()
+		if err := convert(childKey, &keyVal, key.Interface()); err != nil {
+			return errors.Wrap(err, "key types conversion")
+		}
+
+		destMap.SetMapIndex(keyVal, itemValue)
+	}
+
+	value.Set(destMap)
 	return nil
 }
 
@@ -472,7 +529,8 @@ func jsonMap(v interface{}) interface{} {
 	return v
 }
 
-// tryUnmarshallers checks if the value's type implements either one of standard interfaces in order:
+// tryUnmarshallers checks if the value's type implements either one of standard
+// interfaces in order:
 // 1. `json.Unmarshaler`
 // 2. `encoding.TextUnmarshaler`
 // 3. `yaml.Unmarshaler`
@@ -480,15 +538,10 @@ func jsonMap(v interface{}) interface{} {
 func (d *decoder) tryUnmarshalers(key string, value reflect.Value, def string) (bool, error) {
 	switch value.Kind() {
 	// value.IsNil panics if value.Kind() is not equal to one of these constants.
-	case reflect.Chan, reflect.Interface, reflect.Func, reflect.Map, reflect.Ptr:
+	case reflect.Chan, reflect.Interface, reflect.Func, reflect.Map, reflect.Ptr, reflect.Slice:
 		if value.IsNil() {
 			value.Set(reflect.New(value.Type()).Elem())
 		}
-
-	//TODO(alsam) Command-line provider implementation is made via maps, so we'll fail to merge slices and maps
-	// TestCommandLineProvider_NestedValues will fail in that case.
-	case reflect.Slice:
-		return false, nil
 	}
 
 	v := d.getGlobalProvider().Get(key)